MDL-NOBUG session: make acquire timeout "retryable"
When running Moodle in a configuration where FastCGI servers operate
independently of web servers it's necessary to distinguish between
errors that indicate a health problem with a backend and ordinary
operational behaviours.

The Moodle behaviour of raising a 503 where it's not possible to acquire
the lock on the user session within the timeout period can cause
downstream servers to mark servers as unhealthy.

Introduce a separate retryable_exception and catch it during
\core\session\manager::start(), emitting the exception message and
serving a 200 OK rather than 503 Service Unavailable.